chore: restructure code - dashboard views - statistics
